pyproject.toml
[tool.poetry]
name = "acowi-multi-account"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
stripe = "^2.48.0"
python-dotenv = "^0.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
